Enforce Iceberg timestamp precision in ConstraintExtractor
The ConstraintExtractor class has been updated to assert that the precision of TimestampWithTimeZoneType is always 6 in line with Iceberg's limitations. This assertion is done in several methods that handle distinct 'Call' expressions. The assertion is done using the checkArgument method and will cause an error if the condition isn't met.