Add ConstraintExtractor class for Kafka plugin to extract and manipulate constraint
This commit introduces the ConstraintExtractor class in the Kafka connector.

This class is used to extract and manipulate constraint related data and provides static functions for different operations.

It is primarily used to handle different operations such as transforming constraints and expressions, managing conjuncts and converting and intersecting tuple domains.

Moreover, it also handles specific cases such as timestamp casting， truncation and other function(year).